Relocated hub data endpoint unit tests
